+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # Current
 - Added overrides for every() as it fails fast and being able to search in reverse maybe optimal.
 - Made equalsFor and equalsAndHash methods on the hashmap using Object.defineProperty(), in preparation for typed maps.
+- made emplace match the proposal more closely, by not using insert to update.
 - removed jshint and replaced with eslint as the linting tool.
 - More Documentation
 - More Examples